\name{ModuleAApp}
\alias{ModuleAApp}
\title{
Graphical user interface to design an adaptive trial with data-driven sample size or event count re-estimation
}
\description{
This function creates a Shiny-based graphical user interface to facilitate the process of designing Phase III adaptive trials with two interim analysis to enable a futility assessment and sample size/event count re-estimation. The application supports the same functionality as the \code{\link{ModuleA}} function. For an example of the function call, see \link{ExampleA1}.
}
\usage{
ModuleAApp()
}
\seealso{
  \code{\link{ModuleA}}, \code{\link{ModuleAReport}}
}
