\name{ModuleBApp}
\alias{ModuleBApp}
\title{
Graphical user interface to design an adaptive trial with data-driven treatment selection
}
\description{
This function creates a web application with a Shiny-based graphical user interface to facilitate the process of designing Phase III adaptive trials with two interim analysis to enable a futility assessment and data-driven selection of the best treatment arm. The application supports the same functionality as the \code{\link{ModuleB}} function. For an example of the function call, see \link{ExampleB1}.
}
\usage{
ModuleBApp()
}
\seealso{
  \code{\link{ModuleB}}, \code{\link{ModuleBReport}}
}
