\name{ModuleCApp}
\alias{ModuleCApp}
\title{
Graphical user interface to design an adaptive trial with data-driven population selection
}
\description{
This function creates a web application with a Shiny-based graphical user interface to facilitate the process of designing Phase III adaptive trials with two interim analysis to enable a futility assessment and data-driven selection of the most promising patient population or populations. The application supports the same functionality as the \code{\link{ModuleC}} function. For an example of the function call, see \link{ExampleC1}.
}
\usage{
ModuleCApp()
}
\seealso{
  \code{\link{ModuleC}}, \code{\link{ModuleCReport}}
}
