\name{ModuleDApp}
\alias{ModuleDApp}
\title{
Graphical user interface for an optimal selection of a futility stopping rule at an interim analysis
}
\description{
This function creates a web application with a Shiny-based graphical user interface to facilitate the process of defining an optimal futility stopping rule at an interim analysis in a Phase II or Phase III clinical trial. The application supports the same functionality as the \code{\link{ModuleD}} function. For an example of the function call, see \link{ExampleD1}.
}
\usage{
ModuleDApp()
}
\seealso{
  \code{\link{ModuleD}}, \code{\link{ModuleDReport}}
}
